Refactor footer into multiple components
