+++ src/platform/site-wide/index.js
@@ -59,7 +59,7 @@ export default function startSitewideComponents(commonStore) {
   if (brandConsolidation.isEnabled()) {
     startMegaMenuWidget(commonStore);
     startMobileMenuButton(commonStore);
-    startVAFooter(commonStore);
+    startVAFooter();
   }
 
   startMetrics();
