+++ src/platform/site-wide/sass/brand-consolidation.scss
@@ -7,7 +7,7 @@ $font-path: "~uswds/src/fonts";
 $image-path: "~uswds/src/img";
 @import "~@department-of-veterans-affairs/formation/sass/core";
 
-@import "modules/m-footer";
+@import "modules/m-footer-brand-consolidation";
 
 // ----- VA MODULES (BUTTONS/CARDS/TABLES) ---- //
 @import "~@department-of-veterans-affairs/formation/sass/modules/m-alert";
