+++ src/platform/site-wide/sass/modules/_m-footer-brand-consolidation.scss
@@ -1,0 +1,187 @@
+// Footer
+.footer {
+  color: $color-white;
+  margin: 0 auto;
+  background: $color-primary-darkest;
+  position: relative;
+  overflow: hidden;
+
+  @include media($medium-large-screen) {
+    flex-wrap: nowrap;
+  }
+
+  .va-flex {
+    flex-wrap: wrap;
+    padding-left: 0.9375rem;
+    @include media($medium-large-screen) {
+      flex-wrap: nowrap;
+    }
+  }
+}
+
+.footer-inner {
+  padding: 0 0 2rem 0;
+
+  @include media($medium-screen) {
+    padding: 2rem 1rem 0;
+  }
+}
+
+[class^=va-footer-linkgroup] {
+  padding-left: 0;
+  li {
+    list-style: none;
+  }
+
+  a {
+    background: none;
+    color: $color-white;
+    font-size: 1em;
+    font-weight: 400;
+    margin: .25em 0;
+    text-decoration: underline;
+
+    &:hover {
+      color: $color-gold;
+    }
+
+    @include no-sr-content;
+  }
+}
+
+.va-footer-linkgroup {
+  margin: 2em 1em 0 0;
+  flex: 0 1 100%;
+  &:last-child {
+    margin-right: 0;
+  }
+  &:first-child {
+    margin-top: 2em;
+  }
+
+  @include media($medium-large-screen) {
+    flex: 0 1 33%;
+    margin: 2em 1em 1.5em 0;
+  }
+
+  a {
+    display: inline-block;
+  }
+
+  .va-button-link {
+    color: $color-white !important;
+    &:hover {
+      color: $color-gold !important;
+    }
+  }
+}
+
+.va-footer-linkgroup-title {
+  -webkit-font-smoothing: antialiased;
+  color: $color-white;
+  font-weight: 700;
+  font-size: 1.1em;
+  margin: 0;
+  padding: 0 0 1em 0;
+}
+
+.usa-accordion.va-footer-accordion {
+  > li {
+    background: transparent;
+    margin-bottom: 6px !important;
+  }
+}
+
+.va-footer-accordion-content {
+  background: transparent;
+  > .va-footer-links {
+    > li {
+      margin-bottom: 0.5em;
+      > a {
+        display: inline-block;
+        color: $color-white;
+        text-decoration: none;
+        font-size: 1em;
+        font-weight: 400;
+        margin: .25em 0;
+      }
+    }
+  }
+}
+
+.va-footer-links {
+  padding-left: 0;
+  margin-top: 0;
+  > li {
+    margin-bottom: 0;
+  }
+}
+
+.footer-banner,
+.footer-links {
+  padding-left: .5em;
+  padding-right: .5em;
+}
+
+.footer-banner {
+  padding-top: 1.5em;
+  padding-bottom: 1.5em;
+}
+
+.va-footer-button {
+  font-weight: 400;
+  border-top: 3px solid transparent;
+  border-bottom: 3px solid transparent;
+  color: $color-white;
+  cursor: pointer;
+  display: block;
+  margin: 0;
+  padding: 1.6rem 0.8rem;
+  font-size: 1.6rem;
+  font-family: "Source Sans Pro", "Helvetica Neue", "Helvetica", "Roboto", "Arial", sans-serif;
+  color: $color-white;
+  background-color: $color-primary-darker;
+  &:hover {
+    background-color: $color-primary-darker;
+    color: $color-white;
+    border-top: 3px solid transparent;
+    border-bottom: 3px solid transparent;
+  }
+
+  &.usa-accordion-button[aria-expanded=false] {
+    background-image: url("/img/plus-white.png");
+    background-image: url("/img/plus-white.svg");
+  }
+
+  &.usa-accordion-button[aria-expanded=true] {
+    background-image: url("/img/minus-white.png");
+    background-image: url("/img/minus-white.svg");
+  }
+}
+
+.va-footer-links-bottom {
+  ul {
+    padding-left: 0;
+    padding-bottom: 3em;
+    list-style: none;
+
+    li {
+      display: inline-block;
+
+      a {
+        color: $color-white;
+        &:visited {
+          color: $color-white;
+        }
+      }
+      &:after {
+        content: "|";
+        padding: 0 1rem;
+      }
+    }
+  }
+}
+
+.va-footer-content {
+  @include accessibly-hidden();
+}
